Combine wallets into assets card
